requirements.txt
greenlet==1.1.2
mysql-connector-python==8.0.26
numpy==1.21.2
pandas==1.3.3
protobuf==3.18.1
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
python-dateutil==2.8.2
python-decouple==3.5
pytz==2021.3
six==1.16.0
SQLAlchemy==1.4.25